requirements.txt
PyQt6>=6.4.0
python-pptx>=0.6.21
Pillow>=9.0.0
yt-dlp>=2023.1.1
requests>=2.28.0
lxml>=4.9.0
openpyxl>=3.1.0
pywin32>=306

# PowerPoint features (Windows):
# - Requires Microsoft PowerPoint installed
# - Uses VBScript for merging (cscript.exe built into Windows)
# - Uses COM automation via pywin32 for slide thumbnails
